pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "typedb-jupyter"
version = "0.2"
description = "Jupyter connector for TypeDB"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
authors = [
    {name = "James Whiteside", email = "james.whiteside@vaticle.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: IPython",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Topic :: Database",
]
dependencies = [
    "typedb-client>=2.17",
    "ipython"
]

[project.urls]
"Repository" = "https://github.com/james-whiteside/typedb-jupyter"
"TypeDB" = "https://github.com/vaticle/typedb"
"Vaticle" = "https://vaticle.com/"
